+++ web/apps/wps-web/src/features/landingPage/LandingPagePreview/LandingPagePreview.test.tsx
@@ -4,7 +4,7 @@ import { MS_TEAMS_SPRINT_REVIEW_URL } from '@wps/utils/env'
 import { fbpGoInfo, percentileCalcInfo, toolInfos, weatherToolkitInfo } from 'features/landingPage/toolInfo'
 import { MemoryRouter } from 'react-router-dom'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
-import LandingPageRework, { LANDING_PAGE_FAVOURITES_STORAGE_KEY } from './LandingPageRework'
+import LandingPagePreview, { LANDING_PAGE_FAVOURITES_STORAGE_KEY } from './LandingPagePreview'
 
 vi.mock('@wps/utils/env', async importOriginal => {
   const actual = await importOriginal<typeof import('@wps/utils/env')>()
@@ -17,11 +17,11 @@ vi.mock('@wps/utils/env', async importOriginal => {
 const renderPage = () =>
   render(
     <MemoryRouter>
-      <LandingPageRework />
+      <LandingPagePreview />
     </MemoryRouter>
   )
 
-describe('LandingPageRework', () => {
+describe('LandingPagePreview', () => {
   beforeEach(() => {
     localStorage.clear()
   })
