+++ web/apps/wps-web/src/features/landingPage/LandingPagePreview/favouritesStorage.ts
@@ -1,0 +1,28 @@
+import { toolInfos } from 'features/landingPage/toolInfo'
+import { LANDING_PAGE_FAVOURITES_STORAGE_KEY } from './landingPagePreviewConfig'
+
+export const readFavouriteRoutes = () => {
+  try {
+    const storedValue = localStorage.getItem(LANDING_PAGE_FAVOURITES_STORAGE_KEY)
+    if (!storedValue) {
+      return []
+    }
+
+    const parsedValue: unknown = JSON.parse(storedValue)
+    if (!Array.isArray(parsedValue)) {
+      return []
+    }
+
+    return toolInfos.filter(tool => parsedValue.includes(tool.route)).map(tool => tool.route)
+  } catch {
+    return []
+  }
+}
+
+export const storeFavouriteRoutes = (favouriteRoutes: string[]) => {
+  try {
+    localStorage.setItem(LANDING_PAGE_FAVOURITES_STORAGE_KEY, JSON.stringify(favouriteRoutes))
+  } catch {
+    return
+  }
+}
